+++ Makefile
@@ -1,34 +1,22 @@
-install: dependencies build
+.PHONY: build test
 
-build:
-	tar -xf binutils-2.28.1.tar.gz
-	tar -xf gcc-9.3.0-tar.gz
+INCLUDE_DIR=include
+OBJS = $(wildcard $(INCLUDE_DIR)/Build/*.o)
+SOURCE = $(wildcard $(INCLUDE_DIR)/C/*.c)
 
-	export PREFIX="$HOME/opt/cross-test"
-	export TARGET=i686-elf
-	export PATH="$PREFIX/bin:$PATH"
+build: Ghost.bin
 
-	mkdir build-binutils
-	cd build-binutils
-	../binutils-2.28.1/configure --target=$TARGET --prefix="$PREFIX" --with-sysroot --disable-nls --disable-werror
-	make
-	make install
+test:
+	echo $(SOURCE)
 
-	cd ..
+Ghost.bin: kernel.o boot.o $(INCLUDE_DIR)/Build/memory.o linker.ld
+	i686-elf-gcc -T linker.ld -o myos.bin -ffreestanding -O2 -nostdlib boot.o kernel.o $(OBJS) -lgcc
 
-	mkdir build-gcc
-	cd build-gcc
-	../gcc-9.3.0/configure --target=$TARGET --prefix="$PREFIX" --disable-nls --enable-languages=c,c++ --without-headers
-	make all-gcc
-	make all-target-libgcc
-	make install-gcc
-	make install-target-libgcc
+#kernel.o: kernel.c
+#	i686-elf-gcc -c kernel.c -o kernel.o -I$(INCLUDE_DIR)headers -std=gnu99 -ffreestanding -O2 -Wall -Wextra
 
-dependencies:
-	sudo apt install -y build-essential
-	sudo apt install -y bison
-	sudo apt install -y flex
-	sudo apt install -y libgmp3-dev
-	sudo apt install -y libmpc-dev
-	sudo apt install -y libmpfr-dev
-	sudo apt install -y texinfo
+boot.o: boot.s
+	i686-elf-as boot.s -o boot.o
+
+.c.o:
+	i686-elf-gcc -c $^ -o $@ -I$(INCLUDE_DIR)/headers -std=gnu99 -ffreestanding -O2 -Wall -Wextra
